refactor(views): migrate genres to per-folder layout (commit 2/7)
Split src/views/genres.rs (1098 lines) into src/views/genres/{mod,update,view}.rs
following the proven views/settings/ template:

- mod.rs (273): types (Page, Message, Action, ViewData), trait impls
  (HasCommonAction via macro, ViewPage explicit), Default + new(), tests
  for column declarations
- update.rs (339): impl GenresPage { fn update, fn resolve_artwork_action }
  + the toggle-column-visibility test
- view.rs (470): impl GenresPage { fn view, fn render_genre_row,
  fn render_album_row }

First consumer of commit 1's scaffolding macros:
- define_view_columns! replaces the 50-line GenresColumn / Visibility /
  Default / get / set boilerplate with a 7-line declarative table.
- impl_has_common_action!(GenresAction { NavigateAndExpandArtist,
  NavigateAndExpandAlbum }) replaces the 22-line match-arm impl.

No behavioral change. Public API surface unchanged (views/mod.rs
re-exports stay valid). cargo build / clippy --all-targets -D warnings
/ nightly fmt --check / test --workspace all clean (662 + 368 + 1 doctest).
